requirements.txt
streamlit==1.28.0
pandas==2.0.0
plotly==5.15.0
sqlalchemy==2.0.0
psycopg2-binary==2.9.0
pyyaml==6.0